Close #3 with a crude workaround.
AppVeyor was failing because the environment variable holding the rest of the commit message lines after the first one, was missing, when there weren't any, instead of being empty.
Consequently, the variable holding them was null, and you know what happens after that.
I have to make a pull request to FAKE for that.
And fix some recurring Intellisense problems with the FAKE script.