Fix the broken RangeMap.ofSeqEx.
It wasn't actually compacting the consecutive keys; the killer feature of the range map, becoming in essence a glorified array.
This serious bug was first introduced in 6be6570dec3a4ccf8d22a3d686ed6932ea2543c0, in late July 2020, because of a missing single quote from a name.
It was fixed, the variable names became more clear, the implementation was overall optimized, and a test was added to verify the RangeMap's correct behavior.