Precompilables are no longer designtime Farkles.
Special functions were created to build them that fail if the grammar has not been precompiled.
Despite making precompilation less transparent, this change has actually many advantages.
The requirement to mark only the topmost designtime Farkle as precompilable and at the very end is enforced by the compiler.
Furthermore, the IL linker can trim the builder's implementation now that precompiled grammars are not built as a fallback.
As a disadvantage, merely marking a designtime Farkle as precompilable implies a commitment to use the precompiler.
But on the flip side, those using it expect quick startup times and Farkle will not break such expectation.