Define the curly-bracketed regex string quantifiers as terminals.
It allows decimal digits to be freely used in string regexes, but prohibits whitespace inside the curly braces.
That's a minor breaking change I am OK with; it was non-standard behavior.
I could still allow it by explicitly allowing spaces and trimming the span, but there's no reason for such complexity.
And add some tests and update the release notes.