Refactor Position.Advance to correctly handle line endings.
The existing logic was flawed because it discarded blank LF newlines.
With this commit, the (now public) span-based Advance will check if the character before an LF is a CR, instead of checking if the position was at the start of a line.
As a side-effect, Position.Advance and CharStream.AdvanceBy should pass the entire CRLF line ending at once.
And calling `CharStream.AdvanceBy 15` is no more the same with calling `AdvanceBy 1` 15 times.
The release notes were updated.