Remove a strong reference that inhibited the precompiler's ALC from being unloaded.
Woah, that has been these almost since the beginning of the precompiler.
The moral of the story is be very careful with the data you store on an ALC. When it is about to be unloaded, it is held by the runtime, until all its types and instances are collected.
We hold however the input Assembly object inside the ALC object, creating a half-managed and half-native reference cycle which cannot be broken by the GC. Therefore the ALC does not get unloaded.
Even though the precompiler loads the file from memory without locking the file, this commit might fix related user-reported bugs.