Introduce an untyped IPostProcessor interface that has both Transform and Fuse.
It aligns with the rest of Farkle's design philosophy where the generic parameter is an easily decoupled facade, and allows some internal clean-up.
This interface will nowhere be accepted itself in public APIs; custom post-processor implementations will have to implement IPostProcessor<T>, which inherits this interface and has no members (akin to the relationship between the typed and untyped DesigntimeFarkle).