Support any .NET versions later than 6 in the precompiler worker.
Ah, it was so easy! This fixes a major usability roadblock for those using the precompiler with Visual Studio.
It would be the one issue I would backport to 6.x if I could fix, but turns out I don't have to; it can be worked around with an MSBuild script that sets the `DOTNET_ROLL_FORWARD` environment variable to `LatestMajor`.
I can publish a small package to GitHub Packages with just the script.
Also update the MSBuild integration tests to run on .NET 7 to test this.