Check for duplicate accepts in the DFA builder in debug mode only.
I was very torn on this one and originally thought of entirely removing it.
A properly written consumer of this class should never add the same accept many times, so the value of this validation hinges on whether raw creation of state machines is something we want to expose as a public API.
And likely we don't.