Remove `SemanticValue`.
And I liked this idea so much. 😭
The problem with it is that it restricts the API by prohibiting upcasting operations on parser and builder objects, and has a questionable advantage (zero allocations with the calculator)?
Thankfully I just thought of a way to reintroduce an optimization like this in the future.