Parser API design (#89)
* Start writing a design document for the parser API.

* Add specification for the foundational parser APIs.

* Add `ParserState.Context`.

This will allow accessing the context object from the `ParserState` instead of the `ParserInputReader`. The latter's `StateBox` property was removed.

* Add a section with the semantic analysis APIs.

* Move the parser types to the `Farkle.Parser` namespace.

* Specify the parser services that will come with Farkle.

* Remove mentions of pooled buffers in parser state contexts.

The context can automatically return the character buffers all characters have been read and input has completed.

* Add sections with the tokenizer and high-level parser classes.

* Remove the untyped parser interfaces.

They can put a strain on the implementers and are not strictly necessary.
Farkle 6 has a `Cast()` method for runtime Farkles which we might port.

* Remove `SemanticValue`.

And I liked this idea so much. 😭
The problem with it is that it restricts the API by prohibiting upcasting operations on parser and builder objects, and has a questionable advantage (zero allocations with the calculator)?
Thankfully I just thought of a way to reintroduce an optimization like this in the future.

* Rename `Token` to `TokenizerResult` and make it able to encode either tokenizer success or failure.

* Update the parser state's properties.

Add `InputName` and make it and the `Context` get/set instead of get/init; you cannot init the parser state of a `ParserStateContext`.

* Close the `CharParser` API from extensibility.

* Speciffy the top-level parser APIs.

* Move the `ParserResult` types in the root `Farkle` namespace.

* Small edits.

* Make `ParserState.Context` an init-only property again.

It's not supposed to be modified by user code and `ParserStateContext` will still be able to set it to itself.

* Remove a vestige of the untyped parser API.

* Add a note about exception-safety.

* Add a section with some alternative designs for the Parser API.

* Small changes.