Simplify emitting failure transitions.
We cannot skip creating a new state if the state ID is empty; this is incorrect in some edge cases. Instead we emit the failure if we are inside only all inverted ranges.
As part of this the Any leaf type was removed and Any leaves are represented as inverted character leaves with no ranges.