Add a function to enumerate the conflicts of a grammar.
Figuring out a place to put reporting conflicts was hard, with the candidates of doing it at the time of resolving the conflicts or at the time of writing them to the grammar getting rejected for various reasons.
Instead we will do it later, after the grammar gets created by enumerating the conflicts from the grammar's data.