Refactor the nonterminal classes to reduce duplication.
We fix a bug where we made an invalid upcast from `IProduction<T>` to `IProduction`. Turns out the `CastArray` method was not the one that is type-checked.
Because we will have to unwrap the typed productions into a new array, immutable arrays of productions in public APIs were updated to read-only spans.
This early type erasure allows us to update typed nonterminals to contain just an untyped nonterminal to delegate the production setting logic, which now exists in only one place.