Improve the usability of the grammars API et al. (#327)
* Update the `Name` properties of grammar objects to return strings.

* Update properties of grammar objects to return concrete objects instead of handles.

Exceptions are made when the type of object to return is not certain (such as production members).

* Update state machine APIs to return concrete objects instead of handles.

Exceptions are made for LR action types that remain as lightweight as possible (internally backed by a single int).

* Cache grammar string instances.

This eliminates allocations on every `.Name` call and allows us to make string handles internal.
We can always bring them back, but only if there's a good reason.

* Update the grammar format specification to exclude custom streams from inhibiting the parsability of a grammar.

* Devirtualize the `Grammar` property of state machines.

* Use state machine type without conflicts in the parser.
We can do it now that the state machines are not generic, and because they are sealed it devirtualizes all method calls to them.

* Fix tests.

* Fix SonarCloud warning.