Update grammar creation APIs.
We use `Load` instead of `Create` and `Convert` for GOLD Parser grammar conversions.
The read-only span overload was removed because it always had to copy the span if the grammar is valid. Better make memory management explicit to the user.
Also added an API to convert a GOLD Parser grammar from a file.