Various improvements (#331)
* Support equality semantics for concrete grammar objects.

The original plan was to do it only for handles, but since we almost everywhere use concrete objects, they should support equality as well.

* Improve documentation.

* Remove `GrammarExtensions.WriteGrammarToFile`.

It can be done with one line since .NET 9 and with a few more lines in earlier frameworks.

* Make `IProduction` public and have `IProduction<T>` inherit it.

This allows us to accept params immutable array when setting typed productions.

* Fail when setting an empty array of typed productions.

* Refactor `DefaultParserImplementation` to reduce generic instantiations.

It no longer depends on the result type at all, and the APIs that don't even depend on the character type are moved to a non-generic class.

* Fix some Sonar warnings.

* Exclude several diagnostics-only APIs from code coverage.

* Fix F# API.

* Update grammar creation APIs.

We use `Load` instead of `Create` and `Convert` for GOLD Parser grammar conversions.
The read-only span overload was removed because it always had to copy the span if the grammar is valid. Better make memory management explicit to the user.
Also added an API to convert a GOLD Parser grammar from a file.

* Add grammar object equality test.

* Add documentation.