Port templating to FarkleNeo. (#338)
* Port `Farkle.Tools.Shared` to FarkleNeo.

* Remove support for generating grammar skeletons.

Since the introduction of the builder, there's not any use case for them, and it allows us to remove lots of code.

* Fix typo.

* Port the `farkle new` CLI command to FarkleNeo.

* Port the HTML templates to FarkleNeo.

* Remove unused code.

* Support setting the `Generated` flag in the builder.

Currently is it set only for the placeholder nonterminal.

* Display token symbol and nonterminal flags in HTML.

* Add some basic tests for the HTML generator.

* Always emit a navigation bar.

* Make the F# API compatible with nullable reference types.

* Fix tests.

This open was not unused; it brought the auto-opened `Generators` module with the special test functions and Ionide mistakenly suggested to remove it.