Add support for custom DFA start states for groups. (#447)
* Fix grammar table section headers.

* Add specification for DFA group starting states.

* Add public `GroupHandle` type.

* Consistently use the term "start state" over "initial" or "starting" state.

* Support DFA group start states in the grammars API.

* Refactor DFA initialization to pass less parameters and reduce duplication.

* Support DFA group start states in the tokenizer.

* Refactor `DfaWriter` to not require knowing the number of states upfront.

* Add the group parsing algorithm to the grammar format spec.

* Update the group parsing algorithm with what we are going to do.

* Use a dedicated type for the result of DFA matching.

We explicitly store whether the DFA stopped because it accepted or failed, or because it needed more characters to make a conclusion. In the latter case, we return the state at the last accept.

* Update tokenizer implementation according to the spec.

* Bring back `bool isNoise`.

* Simplify checking for void regexes.
We now warn only if the whole regex of a symbol is equivalent to void, by checking if none of the DFA's states accepts the symbol. This lets us move the checks out of the DFA builder.
Also do some clean-up.

* Refactor DFA builder to embed the accept symbols in the regexes.

This increases flexibility, and lets us build regexes that are not simply an alternation of all possible symbols.
Checking for void regexes was moved back to the DFA builder proper.
Also, `Build` became an instance method, and some parameters were moved to the constructor.

* Write the DFA to a user-provided `DfaWriter`.

* Use `GroupHandle` in the grammars writer.

* Use flag enums instead of boolean parameters in the DFA builder.

* Add support for building DFAs that do lazy matching, i.e. stop as soon as they accept a symbol.

* Add high-priority inverted `Chars` regexes.

Also refactor the `AllButChars` kind into a flag.

* Add break-on-accept `Chars` regexes.

* Remove `DfaBuildOptions.LazyMatching`.

It was originally introduced as a much simpler alternative to break-on-accept, but it was later realized that it's not a suitable.
Subsequently, the code and the semantics were simplified, making break-on-accept a viable option.

* Fix state numbering when building multiple DFAs on the same writer.

* Write custom DFA group start states in the builder.

* Add debug browsing support for DFA group start states.

* Fix bugs.

* Remove unnecessary sentence.

Matching nested group start symbols from a custom DFA start state should not be done in general.

* Test tokenizer suspension with group start states.

* Add benchmark of group parsing, with and without custom DFA start states.

* Add new test case that parses from the built GML grammar in dynamic block mode.

The existing test case does not work, because it uses an old GOLD Parser grammar file.

* Update release notes.

* Fix some SonarQube Cloud warnings.

* Improve debugger display of regex flags.