Add native support for individual character set regexes. (#472)
Previously, calling `Regex.OneOf` with a sequence of characters, was converting it to ranges. Now, we reduce upfront computations, by storing the character sequence in the regex, and teaching the DFA builder.
The character set gets converted to range either on canonicalization, or during `BuildDfaStates` if it is already canonicalized.
Some utility code was updated as needed.