Support precompiling assemblies that depend on packages that publish reference assemblies.
This is a known issue for quite some time. The `@(ReferencePath)` item contains assemblies used for compilation, and using it for code execution while filtering out reference assemblies is not valid.
We instead create a separate list of assemblies, and pass it to the task. The previous set of references is still being used by the weaver.